Extract ImageMaid log-content analysis into modules/logscan_imagemaid_analysis.py
Third sub-phase of the logscan/trends breakup. Moves the ImageMaid log
parsing layer: runtime/byte-size parsing, command-snapshot extraction
(used to infer which saved config produced a given run), error-line
extraction, the ~460-line analyze_imagemaid_log_content parser that
turns a raw ImageMaid log into a structured summary, and the progress
snapshot/recommendation builders built on top of it.

Depends only on modules/logscan_resume.py (sub-phase 2, for duration/
byte formatting) and modules.database/helpers/logscan - no dependency
on the not-yet-moved cache layer's sibling functions. Zero monkeypatch
coupling found for any function in this set, so no deferred-import
bridges were needed.

_resolve_imagemaid_run_config_name and _analyze_imagemaid_log_content
are the only two re-exports still needed in quickstart.py; everything
else in this layer is either purely internal to the new module now or
was never called from quickstart.py's remaining code.
